JobQueueMPI.jl

Installation

You can install JobQueueMPI.jl using the Julia package manager. From the Julia REPL, type ] to enter the Pkg REPL mode and run:

pkg> add JobQueueMPI

How it works

First, when running a program using MPI, the user has to set the number of processes that will parallelize the computation. One of these processes will be the controller, and the others will be the workers.

We can easily delimit the areas of the code that will be executed only by the controller or the worker.

JobQueueMPI.jl has the following components:

Controller: The controller is responsible for managing the jobs and the workers. It keeps track of the jobs that have been sent and received and sends the jobs to the available workers.

Worker: The worker is responsible for executing the jobs. It receives the jobs from the controller, executes them, and sends the results back to the controller.

API

JobQueueMPI.Controller — Type

Controller

The controller struct is used to manage the workers and the jobs. It keeps track of the workers' status, the job queue, and the pending jobs. It also keeps track of the last job id that was sent to the workers.

JobQueueMPI.Worker — Type

Worker

A worker process.

JobQueueMPI.add_job_to_queue! — Function

add_job_to_queue!(controller::Controller, message::Any)

Add a job to the controller's job queue.

JobQueueMPI.send_jobs_to_any_available_workers — Function

send_jobs_to_any_available_workers(controller::Controller)

Send jobs to any available workers.

JobQueueMPI.send_termination_message — Function

send_termination_message(controller::Controller)

Send a termination message to all workers.

JobQueueMPI.check_for_job_answers — Function

check_for_job_answers(controller::Controller)

Check if any worker has completed a job and return the answer.

JobQueueMPI.send_job_answer_to_controller — Function

send_job_answer_to_controller(worker::Worker, message)

Send a job answer to the controller process.

JobQueueMPI.receive_job — Function

receive_job(worker::Worker)

Receive a job from the controller process.

JobQueueMPI.get_message — Function

get_message(job::AbstractJob)

Get the message from a job.

JobQueueMPI.pmap — Function

pmap(f, jobs, data_defined_in_process = nothing)

Parallel map function that works with MPI. If the function is called in parallel, it will distribute the jobs to the workers and collect the results. If the function is called in serial, it will just map the function to the jobs.

The function f should take one argument, which is the message to be processed. If data_defined_in_process is not nothing, the function f should take two arguments, the first one being data_defined_in_process.

The controller process will return the answer in the same order as the jobs were given. The workers will return nothing.